Drop OpenSSL for rustls
This simplifies onboarding because it drops a built-by-hand dependency
for a regular cargo one. It also required some environment variables,
which are no longer necessary.

It also reduce "bytes still reachable" in valgrind from ~100k to ~10k
when used by the PHP profiler. These bytes appear to mostly be pseudo-
static globals, but there were so many allocations it is difficult to
tell.

I think overall this is a good change, but it does increase the
size of my final PHP executable from 67M to 84M. We need to spend
some time trying to bring that down soon.