Use panic = "abort" strategy
In the short-term, we are not catching any panics before they cross the
FFI boundary. This strategy is what should be used for that behavior.
Additionally, if we pick unwinding, then users of libddprof have to be
aware of whether we pick libunwind or libgcc_eh. By picking abort, we
avoid this for now.

In the long-term, we expect we'll need to catch panics and unwind.