Add pkg-config variant that sets rpath
The `ddprof_ffi_with_rpath.pc` includes the linker flags
`-Wl,-rpath,${libdir}` so that libddprof can be linked to and used in
the directory it gets installed to.

Without this, we would need to set the same flags in the libddprof user
side, or use `LD_LIBRARY_PATH` or some other mean to tell the OS how
to find libddprof.

These two links helped me figure this out:
* https://amir.rachum.com/blog/2016/09/17/shared-libraries/
* https://github.com/mesonbuild/meson/issues/4027

Also included is a change to `ffi-build.sh` so that the macOS dynamic
library is setup to correctly being used with rpath.