Create `strip`d version of shared library during build
By having the build generate both a `strip`d and an unstrip'd version
of the shared library, we make it easier to package a `strip`d libddprof
without needing access to all architectures and OS's used to build
the binaries.

In particular for Ruby, it allows the Ruby packaging process to be
just a straightforward unpacking + uploading to rubygems.org, without
any kind of changes to the libddprof files.

It also makes it easier to validate the resulting libddprof
distribution: it should match the exact files present in the release
page on GitHub.

Finally, there was a bit of discussion if we should ship both
`strip`d and unstrip'd, or only ship the `strip`d version but I like
having both since it makes it easy to package a version with full
debug info if we need to investigate an issue.