Improve tag parsing and Slices
Previously it couldn't handle a tags like:
  env:staging:east
  value

Tag messages on failure have improved.

As part of adding tests, added more trait support to Slice and
CharSlice so they can be used in equivalence statements.

As part of doing that, I realized Slice could implement IntoIterator
and that it cleaned up 3+ places where we call .into_slice().iter().
I also removed unsafe from `into_slice`. Technically, it could still
fail such as if the user creates a slice of length 4 but only has 2
elements, but this is a user error that cannot really be checked.